server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
scope-optimizer>=0.1.1
python-dotenv>=1.0.0
pydantic>=2.0.0
